toqueteando otra vez el examen rank_03

+++ exam_03/broken_gnl/broken_gnl.c
@@ -6,13 +6,11 @@
 /*   By: carmegon <carmegon@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/18 22:25:38 by carmegon          #+#    #+#             */
-/*   Updated: 2026/01/22 10:47:35 by carmegon         ###   ########.fr       */
+/*   Updated: 2026/03/18 19:01:58 by carmegon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "gnl.h"
-#include <stdlib.h>
-#include <unistd.h>
 
 char	*ft_strchr(char *s, int c)
 {
@@ -25,14 +23,14 @@ char	*ft_strchr(char *s, int c)
 		return (NULL);
 }
 
-void	*ft_memcpy(void *dest, const void *src, int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	while (--n > 0)
 		((char *)dest)[n - 1] = ((char *)src)[n - 1];
 	return (dest);
 }
 
-int	ft_strlen(char *s)
+size_t	ft_strlen(char *s)
 {
 	int	res = 0;
 	while (*s)
@@ -43,7 +41,7 @@ int	ft_strlen(char *s)
 	return (res);
 }
 
-int	str_append_mem(char **s1, char *s2, int size2)
+int	str_append_mem(char **s1, char *s2, size_t size2)
 {
 	int	size1 = ft_strlen(*s1);
 	char	*tmp = malloc(size2 + size1 + 1);
@@ -62,7 +60,7 @@ int	str_append_str(char **s1, char *s2)
 	return (str_append_mem(s1, s2, ft_strlen(s2)));
 }
 
-void	*ft_memmove(void *dest, const void *src, int n)
+void	*ft_memmove(void *dest, const void *src, size_t n)
 {
 	if (dest > src)
 		return (ft_memmove(dest, src, n));
